using $PSScriptRoot instad of myinvocation

+++ Scripts/Cleanup.ps1
@@ -44,7 +44,7 @@ Function Get-ScriptDirectory
 ##########################################################################################
 #   Cleanup
 ##########################################################################################
-$Workdir=get-scriptdirectory
+$workdir=$PSScriptRoot
 
 #load LabConfig
 . "$($workdir)\LabConfig.ps1"
@@ -57,33 +57,52 @@ if (!$prefix){
 WriteInfoHighlighted "VMs:"
 Write-Output (get-vm -Name $prefix*).name
 WriteInfoHighlighted "SwitchName:"
-Write-Output (Get-VMSwitch $prefix*).name
+Write-Output (Get-VMSwitch "$($labconfig.prefix)$($LabConfig.SwitchName)").name
 WriteInfo ""
 
 $answer=read-host "This script will remove all VMs and switches starting with $prefix (all above) Are you sure? (type  Y )"
 
 if ($answer -eq "Y"){
-    Stop-VM $prefix'DC' -TurnOff -Force
-    Restore-VMsnapshot -VmName $prefix'DC' -Name Initial -Confirm:$False
-    Start-Sleep 2
-    Remove-VMsnapshot -VMName $prefix'DC' -name Initial -Confirm:$False
-    get-VM $prefix* | Stop-VM -TurnOff -Force
-    get-VM $prefix* | Remove-VM -Force
-    Remove-VMSwitch $prefix* -Force
+    if ((get-vm "$($prefix)DC")){
+        WriteInfo "Turning off $($prefix)DC"
+        Stop-VM "$($prefix)DC" -TurnOff -Force
+        WriteInfo "Restoring snapshot on $($prefix)DC"
+        Restore-VMsnapshot -VmName "$($prefix)DC"  -Name Initial -Confirm:$False
+        Start-Sleep 2
+        WriteInfo "Removing snapshot from $($prefix)DC"
+        Remove-VMsnapshot -VMName "$($prefix)DC" -name Initial -Confirm:$False
+    }
+    if ((get-VM $prefix*)){
+        WriteInfo "Removing all VMs with prefix $prefix"
+        get-VM $prefix* | Stop-VM -TurnOff -Force
+        get-VM $prefix* | Remove-VM -Force
+    }
+    if ((Get-VMSwitch "$($labconfig.prefix)$($LabConfig.SwitchName)")){
+        WriteInfo "Removing vSwitch $($labconfig.prefix)$($LabConfig.SwitchName)"
+        Remove-VMSwitch "$($labconfig.prefix)$($LabConfig.SwitchName)" -Force
+    }
+    
+    # This is only needed if you kill deployment script in middle when it mounts VHD into mountdir. 
+    if ((Get-ChildItem -Path $workdir\temp\mountdir)){
+        &"$workdir\Tools\dism\dism" /Unmount-Image /MountDir:$workdir\temp\mountdir /discard
+    }
 
-    # This is only needed if you kill deployment script in middle when it mounts VHD into mountdir. If mountdir is empty, it will throw warning..
-    &"$workdir\Tools\dism\dism" /Unmount-Image /MountDir:$Workdir\temp\mountdir /discard
+    if ((Get-Item -Path "$workdir\LAB\VMs")){
+        WriteInfo "Removing folder $workdir\LAB\VMs"
+        remove-item "$workdir\LAB\VMs" -Confirm:$False -Recurse
+    }
 
-    remove-item $workdir\LAB\VMs -Confirm:$False -Recurse
-    remove-item $workdir\temp -Confirm:$False -Recurse -ErrorAction SilentlyContinue 
-
+    if ((Get-Item -Path "$workdir\temp")){
+        WriteInfo "Removing folder $workdir\temp"
+        remove-item $workdir\temp -Confirm:$False -Recurse
+    }
     #Unzipping configuration files as VM was removed few lines ago-and it deletes vm configuration... 
     $zipfile= "$workdir\LAB\DC\Virtual Machines.zip"
     $zipoutput="$workdir\LAB\DC\"
 
     Expand-Archive -Path $zipfile -DestinationPath $zipoutput
 
-    WriteSuccess "Press any key to close window ..."
+    WriteSuccess "Job Done! Press any key to close window ..."
     $host.UI.RawUI.ReadKey("NoEcho,IncludeKeyDown") | OUT-NULL
 }
 else {
